Debloat IVideoDriver and IrrlichtDevice includes
