+++ src/client/imagefilters.cpp
@@ -22,6 +22,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include <cassert>
 #include <vector>
 #include <algorithm>
+#include <IVideoDriver.h>
 
 // Simple 2D bitmap class with just the functionality needed here
 class Bitmap {
