+++ src/client/renderingengine.h
@@ -29,6 +29,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "client/render/core.h"
 // include the shadow mapper classes too
 #include "client/shadows/dynamicshadowsrender.h"
+#include <IVideoDriver.h>
 
 #ifdef SERVER
 #error Do not include in server builds
