+++ src/client/shadows/shadowsScreenQuad.cpp
@@ -18,6 +18,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 */
 
 #include "shadowsScreenQuad.h"
+#include <IVideoDriver.h>
 
 shadowScreenQuad::shadowScreenQuad()
 {
