+++ src/gui/touchcontrols.cpp
@@ -34,6 +34,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "gettext.h"
 #include "IGUIStaticText.h"
 #include "IGUIFont.h"
+#include <IrrlichtDevice.h>
 #include <ISceneCollisionManager.h>
 
 #include <iostream>
