+++ src/irrlichttypes_extrabloated.h
@@ -24,7 +24,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #ifndef SERVER
 #include <IMesh.h>
 #include <IImage.h>
-#include <IrrlichtDevice.h>
 #include <IMeshSceneNode.h>
 #include <IDummyTransformationSceneNode.h>
 #include <SMesh.h>
